fix: the firm floored every order twice and bought the cheap half (#1)
26 of 89 live risk decisions died SIZED_TO_ZERO, every one a request under
six shares, and the executed book skewed toward low-priced names.

Two floors, compounding:

  Runner     floor(notional / price)          -> shares
  Risk Off.  floor(shares x stage x regime)   -> approved

The proposed fix was to size in dollars and convert once. That alone changes
nothing: (N x s)/p and (N/p) x s are the same number, and the floor was doing
all the damage. A 5-share request at 0.1875 is 0.9375 either way. Reordering
the arithmetic does not rescue it; not flooring does.

So quantities are fractional end to end, and the only remaining floor is the
one a non-fractionable asset forces at the broker.

The floor was also directional. A $700 name at a $1,000 slot requested one
share, scaled to 0.1875 and was vetoed; an $18 name requested 55, scaled to
10 and traded. The resulting book is not the book the Evaluator measured --
it is the cheap half of it.

Changed:
- Runner sizing returns an exact fractional quantity. A $1,500 name on a
  $1,000 slot is now holdable at 0.667 shares rather than not at all.
- size_intent scales without flooring and takes a reference price, so
  SIZED_TO_ZERO now means "worth under a dollar" rather than "integer
  arithmetic ate it".
- The portfolio gate's bisection ran over integers, which would have
  reinstated the floor one layer down. It is continuous now, quantized DOWN
  to the broker's nine decimals -- which also restores the veto, since
  bisecting a book with no room lands on 1.8e-11 rather than zero and that
  reads as approved.
- Execution Agent asks Alpaca whether each asset is fractionable and caches
  it. A lookup failure answers "no": under-filling is safe and retryable,
  sending a fraction the venue rejects is not.
- Exits sell float(position), not int(position). Once entries are fractional
  every position is, and truncating strands the remainder.

Six tests asserted the floor as intended behaviour and were rewritten to
assert what replaces it, including one whose tolerance comment admitted the
floor made equal-weight dollars unequal by construction. The unfundable-entry
test kept its invariant (an unfunded entry records flat, or next session
emits a phantom sell -- KI-030) and had its trigger repointed, since its old
premise is now false.

1524 tests, ruff clean, mypy --strict clean on 229 files.

+++ src/shrap/risk_compliance/pre_trade_checker_agent.py
@@ -150,7 +150,7 @@ def _downgrade_to_veto(decision_payload: dict[str, Any], reason_code: str, note:
 
 
 def _scale_down(
-    decision_payload: dict[str, Any], quantity: int, reason_code: str, note: str
+    decision_payload: dict[str, Any], quantity: float, reason_code: str, note: str
 ) -> None:
     """Reduce an approved intent's quantity, keeping it approved.
 
@@ -274,7 +274,7 @@ async def _apply_portfolio_gate(
     assessment = await officer.assess(
         ticker=str(decision_payload.get("ticker", "")),
         side=str(intent.get("side", "buy")),
-        quantity=int(decision_payload.get("approved_quantity", 0)),
+        quantity=float(decision_payload.get("approved_quantity", 0)),
         strategy_ids=[str(s) for s in decision_payload.get("strategy_ids", [])],
         regime_label=label,
         regime_band=band,
@@ -292,7 +292,7 @@ async def _apply_portfolio_gate(
             account_id=assessment.account_id,
         )
         return
-    if assessment.approved_quantity < int(decision_payload.get("approved_quantity", 0)):
+    if assessment.approved_quantity < float(decision_payload.get("approved_quantity", 0)):
         _scale_down(
             decision_payload,
             assessment.approved_quantity,
